refactor(react-components): extract Filter out of SearchAndFilter
