+++ packages/react-components/src/organisms/__tests__/SearchAndFilter.test.tsx
@@ -1,6 +1,5 @@
 import React, { ComponentProps } from 'react';
 import { render } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 
 import SearchAndFilter from '../SearchAndFilter';
 import { noop } from '../../utils';
@@ -12,26 +11,15 @@ const props: ComponentProps<typeof SearchAndFilter> = {
   filterOptions: [],
   filterTitle: '',
 };
-it('renders the search controls', () => {
+it('renders the search and filter controls', () => {
   const { getByRole } = render(<SearchAndFilter {...props} />);
   expect(getByRole('searchbox')).toBeVisible();
   expect(getByRole('button')).toBeVisible();
 });
 
-it('Passes query correctly', () => {
+it('passes query correctly', () => {
   const { getByRole } = render(
     <SearchAndFilter {...props} searchQuery="test123" />,
   );
   expect(getByRole('searchbox')).toHaveValue('test123');
-});
-
-it('shows and hides the dropdown menu', () => {
-  const { getByRole, getByText } = render(
-    <SearchAndFilter {...props} filterTitle="Filter Dropdown" />,
-  );
-  const filterButton = getByRole('button');
-  userEvent.click(filterButton);
-  expect(getByText('Filter Dropdown')).toBeVisible();
-  userEvent.click(filterButton);
-  expect(getByText('Filter Dropdown')).not.toBeVisible();
 });
